backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
cffi==2.0.0
click==8.3.3
cryptography==47.0.0
ecdsa==0.19.2
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
httptools==0.7.1
idna==3.13
Mako==1.3.12
MarkupSafe==3.0.3
passlib==1.7.4
psycopg2-binary==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.27
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
